+++ pkg/blockbuilder/config.go
@@ -1,0 +1,144 @@
+package blockbuilder
+
+import (
+	"encoding/json"
+	"flag"
+	"fmt"
+	"os"
+	"time"
+
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
+	"github.com/twmb/franz-go/pkg/kgo"
+
+	"github.com/grafana/mimir/pkg/storage/tsdb"
+)
+
+type Config struct {
+	InstanceID          string             `yaml:"instance_id" doc:"default=<hostname>" category:"advanced"`
+	PartitionAssignment map[string][]int32 `yaml:"partition_assignment" category:"experimental"`
+
+	ConsumeInterval       time.Duration `yaml:"consume_interval"`
+	ConsumeIntervalBuffer time.Duration `yaml:"consume_interval_buffer"`
+	LookbackOnNoCommit    time.Duration `yaml:"lookback_on_no_commit" category:"advanced"`
+
+	Kafka KafkaConfig `yaml:"kafka"`
+
+	// BlocksStorageConfig is defined outside the block-builder config and is injected dynamically.
+	BlocksStorageConfig tsdb.BlocksStorageConfig `yaml:"-"`
+}
+
+func (cfg *Config) RegisterFlags(f *flag.FlagSet, logger log.Logger) {
+	hostname, err := os.Hostname()
+	if err != nil {
+		level.Error(logger).Log("msg", "failed to get hostname", "err", err)
+		os.Exit(1)
+	}
+
+	cfg.Kafka.RegisterFlagsWithPrefix("block-builder.", f)
+
+	f.StringVar(&cfg.InstanceID, "block-builder.instance-id", hostname, "Instance id.")
+	f.Var(newPartitionAssignmentVar(&cfg.PartitionAssignment), "block-builder.partition-assignment", "Static partition assignment. Format is a JSON encoded map[instance-id][]partitions).")
+	f.DurationVar(&cfg.ConsumeInterval, "block-builder.consume-interval", time.Hour, "Interval between consumption cycles.")
+	f.DurationVar(&cfg.ConsumeIntervalBuffer, "block-builder.consume-interval-buffer", 15*time.Minute, "Extra buffer between subsequent consumption cycles. To avoid small blocks the block-builder consumes until the last hour boundary of the consumption interval, plus the buffer.")
+	f.DurationVar(&cfg.LookbackOnNoCommit, "block-builder.lookback-on-no-commit", 12*time.Hour, "How much of the historical records to look back when there is no kafka commit for a partition.")
+}
+
+func (cfg *Config) Validate() error {
+	if err := cfg.Kafka.Validate(); err != nil {
+		return err
+	}
+
+	if len(cfg.PartitionAssignment) == 0 {
+		return fmt.Errorf("partition assignment is required")
+	}
+	if _, ok := cfg.PartitionAssignment[cfg.InstanceID]; !ok {
+		return fmt.Errorf("instance id %q must be present in partition assignment", cfg.InstanceID)
+	}
+	// TODO(codesome): validate the consumption interval. Must be <=2h and can divide 2h into an integer.
+	if cfg.ConsumeInterval < 0 {
+		return fmt.Errorf("consume-interval cannot be negative")
+	}
+	if cfg.LookbackOnNoCommit < 0 {
+		return fmt.Errorf("lookback-on-no-commit cannot be negative")
+	}
+
+	return nil
+}
+
+type KafkaConfig struct {
+	Address       string        `yaml:"address"`
+	Topic         string        `yaml:"topic"`
+	ClientID      string        `yaml:"client_id"`
+	DialTimeout   time.Duration `yaml:"dial_timeout"`
+	ConsumerGroup string        `yaml:"consumer_group"`
+}
+
+func (cfg *KafkaConfig) RegisterFlagsWithPrefix(prefix string, f *flag.FlagSet) {
+	f.StringVar(&cfg.Address, prefix+"kafka.address", "", "The Kafka seed broker address.")
+	f.StringVar(&cfg.Topic, prefix+"kafka.topic", "", "The Kafka topic name.")
+	f.StringVar(&cfg.ClientID, prefix+"kafka.client-id", "", "The Kafka client ID.")
+	f.DurationVar(&cfg.DialTimeout, prefix+"kafka.dial-timeout", 5*time.Second, "The maximum time allowed to open a connection to a Kafka broker.")
+	f.StringVar(&cfg.ConsumerGroup, prefix+"kafka.consumer-group", "block-builder", "The consumer group used to keep track of the consumed offsets for each partition.")
+}
+
+func (cfg *KafkaConfig) Validate() error {
+	return nil
+}
+
+type partitionAssignmentVar map[string][]int32
+
+func newPartitionAssignmentVar(p *map[string][]int32) *partitionAssignmentVar {
+	return (*partitionAssignmentVar)(p)
+}
+
+func (v *partitionAssignmentVar) Set(s string) error {
+	if s == "" {
+		return nil
+	}
+	val := make(map[string][]int32)
+	err := json.Unmarshal([]byte(s), &val)
+	if err != nil {
+		return fmt.Errorf("unmarshal partition assignment: %w", err)
+	}
+	*v = val
+	return nil
+}
+
+func (v partitionAssignmentVar) String() string {
+	return fmt.Sprintf("%v", map[string][]int32(v))
+}
+
+type kafkaLogger struct {
+	logger log.Logger
+}
+
+func newKafkaLogger(logger log.Logger) *kafkaLogger {
+	return &kafkaLogger{
+		logger: log.With(logger, "component", "kafka_client"),
+	}
+}
+
+func (l *kafkaLogger) Level() kgo.LogLevel {
+	// The Kafka client calls Level() to check whether debug level is enabled or not.
+	// To keep it simple, we always return Info, so the Kafka client will never try
+	// to log expensive debug messages.
+	return kgo.LogLevelInfo
+}
+
+func (l *kafkaLogger) Log(lev kgo.LogLevel, msg string, keyvals ...any) {
+	if lev == kgo.LogLevelNone {
+		return
+	}
+	keyvals = append([]any{"msg", msg}, keyvals...)
+	switch lev {
+	case kgo.LogLevelDebug:
+		level.Debug(l.logger).Log(keyvals...)
+	case kgo.LogLevelInfo:
+		level.Info(l.logger).Log(keyvals...)
+	case kgo.LogLevelWarn:
+		level.Warn(l.logger).Log(keyvals...)
+	case kgo.LogLevelError:
+		level.Error(l.logger).Log(keyvals...)
+	}
+}
