+++ pkg/blockbuilder/kafkautil_test.go
@@ -78,20 +78,24 @@ func TestKafkaGetGroupLag(t *testing.T) {
 		require.NoError(t, err)
 
 		require.EqualValues(t, 0, getTopicPartitionLag(t, groupLag, testTopic, 0), "partition 0 must have no lag")
-		require.EqualValues(t, numRecords, getTopicPartitionLag(t, groupLag, testTopic, 1), "partition 1 must fall back to start and get its lag from there")
+		require.EqualValues(t, numRecords, getTopicPartitionLag(t, groupLag, testTopic, 1), "partition 1 must fall back to the start and get its lag from there")
 		require.EqualValues(t, 0, getTopicPartitionLag(t, groupLag, testTopic, 2), "partition 2 has no data and must have no lag")
 	})
 
 	t.Run("fallbackOffset=end", func(t *testing.T) {
-		_, err := getGroupLag(ctx, admClient, testTopic, testGroup, -1)
-		require.Error(t, err, "fallback to partition end isn't supported")
+		groupLag, err := getGroupLag(ctx, admClient, testTopic, testGroup, kafkaOffsetEnd)
+		require.NoError(t, err)
+
+		require.EqualValues(t, 0, getTopicPartitionLag(t, groupLag, testTopic, 0), "partition 0 must have no lag")
+		require.EqualValues(t, 0, getTopicPartitionLag(t, groupLag, testTopic, 1), "partition 1 must fall back to the end and get no lag")
+		require.EqualValues(t, 0, getTopicPartitionLag(t, groupLag, testTopic, 2), "partition 2 has no data and must have no lag")
 	})
 
 	t.Run("group=unknown", func(t *testing.T) {
 		groupLag, err := getGroupLag(ctx, admClient, testTopic, "unknown", kafkaOffsetStart)
 		require.NoError(t, err)
 
-		// This group doesn't have any commits so it must calc its lag from the fallback.
+		// This group doesn't have any commits, so it must calc its lag from the fallback.
 		require.EqualValues(t, numRecords, getTopicPartitionLag(t, groupLag, testTopic, 0))
 		require.EqualValues(t, numRecords, getTopicPartitionLag(t, groupLag, testTopic, 1))
 		require.EqualValues(t, 0, getTopicPartitionLag(t, groupLag, testTopic, 2), "partition 2 has no data and must have no lag")
